Write tests and implement functions being tested
The capitalize function capitalizes the first character in a given
string. The test simply checks to see if the function will return a
new string with just the first character capitalized.

The reverseString function takes in a string and returns it reversed.
The test simply checks to see if the function returns a reversed
version of the input string.

The add, subtract, divide, and multiply methods within a calculator
object are also tested by making sure the methods return the correct
values.